ui-core: fix circular dependency in tolerance picker
Fix the following warning by moving a constant in a shared file:

    (!) Circular dependency
    src/components/inputs/tolerancePicker/TolerancePicker.tsx -> src/components/inputs/tolerancePicker/ToleranceRangeGrid.tsx -> src/components/inputs/tolerancePicker/TolerancePicker.tsx
