+++ ui-core/src/components/inputs/tolerancePicker/TolerancePicker.tsx
@@ -1,32 +1,11 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { useState, useRef, useEffect } from 'react';
 
+import { TOLERANCE_RANGES } from './consts';
 import ToleranceRangeGrid from './ToleranceRangeGrid';
 import Modal from '../../Modal';
 import Input, { type InputProps } from '../Input';
 import { type StatusWithMessage } from '../StatusMessage';
-
-export const TOLERANCE_RANGES = [
-  { label: '00', value: 0 },
-  { label: '05', value: 300 },
-  { label: '10', value: 600 },
-  { label: '15', value: 900 },
-  { label: '20', value: 1200 },
-  { label: '25', value: 1500 },
-  { label: '30', value: 1800 },
-  { label: '35', value: 2100 },
-  { label: '40', value: 2400 },
-  { label: '45', value: 2700 },
-  { label: '60', value: 3600 },
-  { label: '75', value: 4500 },
-  { label: '90', value: 5400 },
-  { label: '105', value: 6300 },
-  { label: '120', value: 7200 },
-  { label: '150', value: 9000 },
-  { label: '180', value: 10800 },
-  { label: '210', value: 12600 },
-  { label: '240', value: 14400 },
-];
 
 export type ToleranceValues = {
   minusTolerance: number;
